mutexes added, bug fixes, new maps, a lot of little tweaks

+++ fdf_other.c
@@ -6,18 +6,18 @@
 /*   By: mperseus <mperseus@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/01/10 03:32:09 by mperseus          #+#    #+#             */
-/*   Updated: 2020/01/10 03:32:30 by mperseus         ###   ########.fr       */
+/*   Updated: 2020/02/15 03:42:04 by mperseus         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "fdf.h"
 
-float	deg_to_rad(int degrees)
+double	deg_to_rad(int degrees)
 {
-	return ((float)degrees * PI / 180.0);
+	return ((double)degrees * PI / 180.0);
 }
 
-int		rad_to_deg(float radians)
+int		rad_to_deg(double radians)
 {
 	return ((int)(round(radians * 180.0 / PI)));
 }
